Split into smaller components for better readability
